backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aioredis>=2.0.1",
    "celery>=5.4.0",
    "diffusers>=0.30.3",
    "fastapi[standard]>=0.115.2",
    "pillow>=11.0.0",
    "pydantic>=2.9.2",
    "python-multipart>=0.0.12",
    "redis>=5.1.1",
    "sqlalchemy>=2.0.36",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "transformers>=4.45.2",
    "uuid>=1.30",
    "uvicorn>=0.32.0",
    "websockets>=13.1",
]
